% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{avg_expr}
\alias{avg_expr}
\title{Average expression of duplicate genes}
\usage{
avg_expr(gene_expr)
}
\arguments{
\item{gene_expr}{gene expression matrix}

\item{dupl_genes}{vector of duplicate genes}
}
\value{
gene expression matrix with averaged expression of duplicate genes
}
\description{
Average expression of duplicate genes
}
\examples{
avg_expr(gene_expr, c("A", "B", "C")))
}
