% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{BFS}
\alias{BFS}
\title{Breadth-first search over a graph}
\usage{
BFS(Adj, root)
}
\arguments{
\item{Adj}{Adjacency matrix.}

\item{root}{Starting vertex index.}
}
\value{
Integer vector of visited vertices.
}
\description{
Breadth-first search over a graph
}
\keyword{internal}
