% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_weights.R
\name{compute_lr_weight}
\alias{compute_lr_weight}
\title{Compute weights for ligand-receptor pairs}
\usage{
compute_lr_weight(i, log2_tpm, lr_pairs)
}
\arguments{
\item{i}{Index of ligand-receptor pair}

\item{log2_tpm}{Log2-transformed TPM matrix}

\item{lr_pairs}{Data frame containing ligand-receptor pairs}
}
\value{
Data frame containing weights for each patient
}
\description{
Patient-specific LR pair weights were defined as the minimum of
log2(TPM+1) expression 613 of the ligand and the receptor, hypothesizing
that the expression of the gene at the lower level 614 limits the LR
binding affinity.
}
\examples{
 dontrun{
compute_lr_weight(1, log2_tpm, lr_pairs)}
}
