% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Count_Types}
\alias{Count_Types}
\title{Count graph motifs by cell-type combination}
\usage{
Count_Types(oblist, V, maxTypes = 0)
}
\arguments{
\item{oblist}{Matrix of graph objects where each row contains vertex indices.}

\item{V}{Integer vector mapping vertices to cell types.}

\item{maxTypes}{Optional maximum number of types used to size the output array.}
}
\value{
An array counting how often each type combination occurs.
}
\description{
Count graph motifs by cell-type combination
}
