% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaCInG_input_generation.R
\name{createCellLigList}
\alias{createCellLigList}
\title{Read a cell-to-ligand compatibility matrix}
\usage{
createCellLigList(filename)
}
\arguments{
\item{filename}{Path to a CSV file with cell types in rows and ligands in columns.}
}
\value{
A list containing the matrix plus ligand and cell-type labels.
}
\description{
Read a cell-to-ligand compatibility matrix
}
