% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{Cycle_Triangles}
\alias{Cycle_Triangles}
\title{Enumerate directed cycle triangles}
\usage{
Cycle_Triangles(Adj)
}
\arguments{
\item{Adj}{Adjacency matrix.}
}
\value{
A list with the cycle-triangle count and the vertex triplets found.
}
\description{
Enumerate directed cycle triangles
}
