% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{detect_duplicate_genes}
\alias{detect_duplicate_genes}
\title{Detect duplicate genes in expression matrix}
\usage{
detect_duplicate_genes(genes)
}
\arguments{
\item{genes}{vector of gene symbols}
}
\value{
vector of duplicate genes
}
\description{
Detect duplicate genes in expression matrix
}
\examples{
detect_duplicate_genes(c("A", "B", "A", "C"))
}
