% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{EdgetoAdj}
\alias{EdgetoAdj}
\title{Convert an edge list to a sparse adjacency matrix}
\usage{
EdgetoAdj(E, N)
}
\arguments{
\item{E}{Two-column edge list with source and target vertex indices.}

\item{N}{Total number of vertices.}
}
\value{
A sparse adjacency matrix of class \code{dgCMatrix}.
}
\description{
Convert an edge list to a sparse adjacency matrix
}
