% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{Find_Number_2Loops}
\alias{Find_Number_2Loops}
\title{Count reciprocal 2-loops allowing multi-edges}
\usage{
Find_Number_2Loops(Adj)
}
\arguments{
\item{Adj}{Adjacency matrix.}
}
\value{
Integer count of 2-loops.
}
\description{
Count reciprocal 2-loops allowing multi-edges
}
