% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{Find_Number_2Loops_Unique}
\alias{Find_Number_2Loops_Unique}
\title{Count unique reciprocal 2-loops}
\usage{
Find_Number_2Loops_Unique(Adj)
}
\arguments{
\item{Adj}{Adjacency matrix.}
}
\value{
Integer count of unique 2-loops.
}
\description{
Count unique reciprocal 2-loops
}
