% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{Find_Number_Triangles}
\alias{Find_Number_Triangles}
\title{Count triangles allowing multi-edges}
\usage{
Find_Number_Triangles(Adj)
}
\arguments{
\item{Adj}{Adjacency matrix.}
}
\value{
Integer count of triangles.
}
\description{
Count triangles allowing multi-edges
}
