% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{Find_Number_Trust_Triangles_Unique}
\alias{Find_Number_Trust_Triangles_Unique}
\title{Count unique trust triangles in a directed graph}
\usage{
Find_Number_Trust_Triangles_Unique(Adj)
}
\arguments{
\item{Adj}{Adjacency matrix.}
}
\value{
Integer count of trust triangles.
}
\description{
Count unique trust triangles in a directed graph
}
