% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{Find_Number_Wedges}
\alias{Find_Number_Wedges}
\title{Count wedges allowing multi-edges}
\usage{
Find_Number_Wedges(Adj)
}
\arguments{
\item{Adj}{Adjacency matrix.}
}
\value{
Integer count of wedges.
}
\description{
Count wedges allowing multi-edges
}
