% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{Find_Number_Wedges_Unique}
\alias{Find_Number_Wedges_Unique}
\title{Count unique wedges in a directed graph}
\usage{
Find_Number_Wedges_Unique(Adj)
}
\arguments{
\item{Adj}{Adjacency matrix.}
}
\value{
Integer count of unique wedges.
}
\description{
Count unique wedges in a directed graph
}
