% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_generation.R
\name{genRandomCellLigands}
\alias{genRandomCellLigands}
\title{Generate a random cell-to-ligand compatibility matrix}
\usage{
genRandomCellLigands(cellTypeNo, ligNo)
}
\arguments{
\item{cellTypeNo}{Number of cell types.}

\item{ligNo}{Number of ligand types.}
}
\value{
A binary matrix describing ligand availability per cell type.
}
\description{
Generate a random cell-to-ligand compatibility matrix
}
