% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_generation.R
\name{genRandomCellReceptors}
\alias{genRandomCellReceptors}
\title{Generate a random cell-to-receptor compatibility matrix}
\usage{
genRandomCellReceptors(cellTypeNo, recNo)
}
\arguments{
\item{cellTypeNo}{Number of cell types.}

\item{recNo}{Number of receptor types.}
}
\value{
A binary matrix describing receptor availability per cell type.
}
\description{
Generate a random cell-to-receptor compatibility matrix
}
