% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_generation.R
\name{genRandomCellTypeDistr}
\alias{genRandomCellTypeDistr}
\title{Generate a random cell-type distribution}
\usage{
genRandomCellTypeDistr(cellTypeNo)
}
\arguments{
\item{cellTypeNo}{Number of cell types.}
}
\value{
A probability vector summing to 1.
}
\description{
Generate a random cell-type distribution
}
