% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_generation.R
\name{genRandomCellTypeList}
\alias{genRandomCellTypeList}
\title{Sample cell-type labels for graph vertices}
\usage{
genRandomCellTypeList(Dcelltype, vertexNo)
}
\arguments{
\item{Dcelltype}{Probability vector over cell types.}

\item{vertexNo}{Number of vertices to sample.}
}
\value{
An integer vector of sampled cell-type indices.
}
\description{
Sample cell-type labels for graph vertices
}
