% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_generation.R
\name{genRandomLigRecDistr}
\alias{genRandomLigRecDistr}
\title{Generate a random ligand-receptor distribution}
\usage{
genRandomLigRecDistr(ligNo, recNo)
}
\arguments{
\item{ligNo}{Number of ligand types.}

\item{recNo}{Number of receptor types.}
}
\value{
A normalized ligand-by-receptor probability matrix.
}
\description{
Generate a random ligand-receptor distribution
}
