% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_name}
\alias{get_name}
\title{Get name of file without extension}
\usage{
get_name(filepath)
}
\arguments{
\item{filepath}{Path to file}
}
\value{
Name of file without extension
}
\description{
Get name of file without extension
}
\examples{
get_name("R/my_script.R") returns 'my_script'
}
