% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{GSCC}
\alias{GSCC}
\title{Extract the giant strongly connected component}
\usage{
GSCC(Adj)
}
\arguments{
\item{Adj}{Adjacency matrix.}
}
\value{
Integer vector of vertex indices in the GSCC.
}
\description{
Extract the giant strongly connected component
}
