% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{IN}
\alias{IN}
\title{Compute the IN component of a directed graph}
\usage{
IN(Adj)
}
\arguments{
\item{Adj}{Adjacency matrix.}
}
\value{
Integer vector of vertices in the IN component.
}
\description{
Compute the IN component of a directed graph
}
