% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{OUT}
\alias{OUT}
\title{Compute the OUT component of a directed graph}
\usage{
OUT(Adj)
}
\arguments{
\item{Adj}{Adjacency matrix.}
}
\value{
Integer vector of vertices in the OUT component.
}
\description{
Compute the OUT component of a directed graph
}
