% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{poiBPFunc}
\alias{poiBPFunc}
\title{Poisson branching-process helper}
\usage{
poiBPFunc(x, M, sens)
}
\arguments{
\item{x}{Current estimate vector.}

\item{M}{Mean offspring matrix.}

\item{sens}{Length of \code{x}.}
}
\value{
A numeric vector used by \code{nleqslv::nleqslv()}.
}
\description{
Poisson branching-process helper
}
\keyword{internal}
