% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{range01}
\alias{range01}
\title{Scale values to \link{0, 1}}
\usage{
range01(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
vector with values scaled to \link{0, 1}
}
\description{
Scale values to \link{0, 1}
}
\examples{
range01(c(1, 2, 3))
}
