% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt_to_csv.R
\name{Read_Sim_Output}
\alias{Read_Sim_Output}
\title{Read a RaCInG simulation output file}
\usage{
Read_Sim_Output(filename)
}
\arguments{
\item{filename}{Path to a \code{.out} file generated by \code{\link[=runSim]{runSim()}}.}
}
\value{
A list containing the average values, standard deviations, and raw counts.
}
\description{
Read a RaCInG simulation output file
}
