% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{remove_dupl_genes}
\alias{remove_dupl_genes}
\title{Remove duplicate genes from expression matrix by averaging their expression}
\usage{
remove_dupl_genes(gene_expr)
}
\arguments{
\item{gene_expr}{gene expression matrix}
}
\value{
gene expression matrix without duplicate genes
}
\description{
Remove duplicate genes from expression matrix by averaging their expression
}
\examples{
remove_dupl_genes(gene_expr)
}
