% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{TarjanIterative}
\alias{TarjanIterative}
\title{Find strongly connected components with Tarjan's algorithm}
\usage{
TarjanIterative(Adj)
}
\arguments{
\item{Adj}{Adjacency matrix.}
}
\value{
A list of strongly connected components.
}
\description{
Find strongly connected components with Tarjan's algorithm
}
\keyword{internal}
