% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{Trust_Triangles}
\alias{Trust_Triangles}
\title{Enumerate outward trust triangles}
\usage{
Trust_Triangles(Adj)
}
\arguments{
\item{Adj}{Adjacency matrix.}
}
\value{
A list with the triangle count and the vertex triplets found.
}
\description{
Enumerate outward trust triangles
}
