% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{Wedges}
\alias{Wedges}
\title{Enumerate wedges in a directed graph}
\usage{
Wedges(Adj)
}
\arguments{
\item{Adj}{Adjacency matrix.}
}
\value{
A list with the wedge count and the vertex triplets found.
}
\description{
Enumerate wedges in a directed graph
}
